examples: Add a 3d example

+++ src/rprocessing/RLangPApplet.java
@@ -214,8 +214,9 @@ public class RLangPApplet extends BuiltinApplet {
         this.renjinEngine.put("displayHeight", displayHeight);
         this.renjinEngine.put("focused", focused);
         this.renjinEngine.put("keyPressed", keyPressed);
-        this.renjinEngine.put("frameCount", frameCount);
+        this.renjinEngine.put("frameCount", (double) frameCount);
         this.renjinEngine.put("frameRate", frameRate);
+        // TODO: Find some ways to push constants into R.
     }
 
     /**
